requirements.txt
asgiref==3.4.1
Django==3.2.6
django-multiselectfield==0.1.12
django-url-params==1.0.2
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
django-heroku==0.3.1
gunicorn==20.1.0
secure-smtplib==0.1.1
python-dateutil==2.8.1